refactor: 사용하지 않는 패키지 제거, 빌드시 eslint 우회
