+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import Background from "../components/Background";
 import Header from "../components/Header";
 import Main from "../components/main/Main";
-import { TmpContextProvider } from "./TmpContext";
+import { TmpContextProvider } from "../components/TmpContext";
 
 export default function Home() {
   return (
